Lift the lesson-synthesis rule into a shared leaf crate
is_action_tool existed twice — arkavo-server synthesis.rs and the shadow
job's episodes.rs — with nothing pinning the copies together. Both now
re-export the single definition from arkavo-lesson-synthesis, a std-only
leaf crate, so the runtime teacher and the audit plane cannot drift on
what counts as an action. Zero-runtime-contact still holds: the shared
crate has no dependencies at all.
