+++ crates/arkavo-lesson-synthesis/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "arkavo-lesson-synthesis"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+authors.workspace = true
+repository.workspace = true
+description = "The lesson-synthesis rules shared by the runtime consolidation teacher and the audit-plane shadow job."
+
+[lints]
+workspace = true
+
+[dependencies]
