+++ crates/arkavo-server/Cargo.toml
@@ -84,6 +84,7 @@ arkavo-tasks = { path = "../arkavo-tasks" }
 arkavo-mcp = { path = "../arkavo-mcp" }
 arkavo-mcp-tools = { path = "../arkavo-mcp-tools" }
 arkavo-memory = { path = "../arkavo-memory" }
+arkavo-lesson-synthesis = { path = "../arkavo-lesson-synthesis" }
 arkavo-llm = { path = "../arkavo-llm", default-features = false, features = ["kimi"] }
 # Direct llama.cpp handle for LlamaTokenEstimator (real tokenizer); only under llama-cpp.
 arkavo-llama-cpp = { path = "../arkavo-llama-cpp", optional = true }
