+++ crates/arkavo-server/src/server/synthesis.rs
@@ -13,18 +13,10 @@ use uuid::Uuid;
 
 use super::episode_buffer::ToolObservation;
 
-/// The lesson-synthesis rule: a tool is an *action* tool only when it is not
-/// a read-only observe/list/hash/summary call. Lessons must reason about
-/// which actions produce good outcomes, not tool-calling mechanics. This is
-/// the single in-crate definition — the consolidation teacher consumes it
-/// too; do not re-inline it elsewhere.
-pub(super) fn is_action_tool(tool: &str) -> bool {
-    let lower = tool.to_lowercase();
-    !lower.contains("observe")
-        && !lower.contains("list")
-        && !lower.contains("hash")
-        && !lower.contains("summary")
-}
+// The lesson-synthesis rule lives in arkavo-lesson-synthesis, shared with
+// the audit-plane shadow job so the two planes cannot drift; the
+// consolidation teacher consumes it through this re-export.
+pub(super) use arkavo_lesson_synthesis::is_action_tool;
 
 /// Truncate a result string to fit within a character budget.
 fn truncate_result(result: &str, max_chars: usize) -> String {
